chore(footnotes): move helper functions into a separate file
